onTransaction event. Ported playerdata fns.
