+++ Assets/Scripts/DataModels/PlayerData.cs
@@ -1,6 +1,4 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
+﻿using UnityEngine;
 
 
 [CreateAssetMenu(fileName = "PlayerData", menuName = "ScriptableObjects/PlayerData", order = 1)]
@@ -8,25 +6,4 @@ public class PlayerData : ScriptableObject
 {
     public long playerMoney;
     public long playerTotalMoneyAcquired; //Used to unlock missions
-
-    public bool CanSpendMoney(long amount)
-    {
-        if (amount < playerMoney)
-        {
-            return true;
-        }
-
-        return false;
-    }
-
-    public void SpendMoney(long amount)
-    {
-        playerMoney -= amount;
-    }
-
-    public void ReceiveMoney(long amount)
-    {
-        playerMoney += amount;
-        playerTotalMoneyAcquired += amount;
-    }
 }
