+++ Assets/Scripts/Singletons/PlayerManager.cs
@@ -13,6 +13,8 @@ public class PlayerManager : MonoBehaviour
     public bool isPaused;
     public PlayerMovement playerMovement;
 
+    public static event System.Action<long> onFinancialTransaction;
+
     void Awake()
     {
         if (Instance == null)
@@ -37,13 +39,26 @@ public class PlayerManager : MonoBehaviour
         }
     }
 
-    void Update()
+    public bool CanSpendMoney(long amount)
     {
-        
+        if (amount < Instance.playerData.playerMoney)
+        {
+            return true;
+        }
+
+        return false;
     }
 
-    public static void VendingMachine()
+    public void SpendMoney(long amount)
     {
+        Instance.playerData.playerMoney -= amount;
+        onFinancialTransaction?.Invoke(Instance.playerData.playerMoney);
+    }
 
+    public void ReceiveMoney(long amount)
+    {
+        Instance.playerData.playerMoney += amount;
+        Instance.playerData.playerTotalMoneyAcquired += amount;
+        onFinancialTransaction?.Invoke(Instance.playerData.playerMoney);
     }
 }
